feat(clightning): use developer build of c-lightning
In production builds of lightningd, it automatically polls the bitcoind node every 30 seconds. This
causes an UX issue because Polar then needs to wait 30 seconds to see the result of any onchain
txns. Using the developer build, we can use the --dev-bitcoind-poll=<seconds> argument to decrease
the polling interval down to 2 seconds