refactor(appdata): store appData in ~/.config instead of APPDATA
Changed the folder where the node docker volumes are mounted on the host

BREAKING CHANGE: With c-lighting changing the folder structure of their dataDir in v0.8.0,
lightningd was unable to create the lightning-rpc socket file due to 'path too long' errors on Mac.
The node data was originally stored in ~/Library/Application Support/polar/data/networks/. This has
now been changed to ~/.polar on all OS's. Due to this change, networks that were created in Polar
v0.1.0 will no longer be displayed in the app. They will need to be mnually moved or re-created.