feat: add import/export network functionality
Add functionality to export and import networks and data from the application. On export we zip the
docker-compose file, the data directories for the nodes and JSON representations of the chart and
network definitions. This zip file can later be imported, where we do the reverse operation and copy
the docker-compose file and the data directories, and insert the network and chart definitions into
the networks.json file.

fix 203