fix(export): fix archiver module not found error when packaged
When packaging the app, it would not load due to an "module 'archiver' not found" error. This was
caused by it being listed as an "external" in the webpack config. That webpack exclusion was
previously needed because the library wouldn't work in chromium without it. The zip/unzip logic has
now been moved to the main renderer process in order to get it working.